Move element registration, scroll-spy and scrolling animation to services
* Moves the major players to services which can be utilized outside of
  the mixins

Closes #10
