+++ modules/index.js
@@ -1,3 +1,5 @@
 exports.Link = require('./components/Link.js');
 exports.Element = require('./components/Element.js');
 exports.Helpers = require('./mixins/Helpers.js');
+exports.scroller = require('./mixins/scroller.js');
+exports.scrollSpy = require('./mixins/scroll-spy.js');
